# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Plane ticket",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    },
    {
      "url": "https://binar.ccrsxx.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user",
        "description": "Get current user",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "200 - Get current user",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "256"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Tue, 29 Oct 2024 04:13:27 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"100-n2PTDtqSVspCElSvg8Dh/TCLs74\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "createdAt": {
                          "type": "string",
                          "example": "2024-10-25T12:51:28.774Z"
                        },
                        "email": {
                          "type": "string",
                          "example": "aminrisal@gmail.com"
                        },
                        "id": {
                          "type": "string",
                          "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                        },
                        "name": {
                          "type": "string",
                          "example": "Risal Amin"
                        },
                        "password": {
                          "type": "string",
                          "example": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2024-10-25T12:51:28.774Z"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "200 - Get current user": {
                    "value": {
                      "data": {
                        "createdAt": "2024-10-25T12:51:28.774Z",
                        "email": "aminrisal@gmail.com",
                        "id": "5ec59a42-64f4-437d-8477-3b3db84c60aa",
                        "name": "Risal Amin",
                        "password": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq",
                        "updatedAt": "2024-10-25T12:51:28.774Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401 - Invalid token",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "41"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Mon, 28 Oct 2024 15:00:10 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"29-ZgbqETaDNxqczg11v3KWsPM+ATo\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid authorization token"
                    }
                  }
                },
                "examples": {
                  "401 - Invalid token": {
                    "value": {
                      "message": "Invalid authorization token"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["User"],
        "summary": "Update current user",
        "description": "Update current user",
        "operationId": "updateCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "Jakarta"
                  },
                  "identityNumber": {
                    "type": "string",
                    "example": "123456789"
                  },
                  "identityType": {
                    "type": "string",
                    "example": "PASSPORT"
                  },
                  "image": {
                    "type": "string",
                    "example": "https://ik.imagekit.io/ccrsxx/c093c949-6dc5-40cd-88a8-e6824e78730a_ysBQ8_ikR.jpg"
                  }
                }
              },
              "examples": {
                "Update current user": {
                  "value": {
                    "address": "Jakarta",
                    "identityNumber": "123456789",
                    "identityType": "PASSPORT",
                    "image": "https://ik.imagekit.io/ccrsxx/c093c949-6dc5-40cd-88a8-e6824e78730a_ysBQ8_ikR.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 - Update current user profile",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "606"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 07 Nov 2024 14:53:40 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"25e-yItjkROA75/0xqLVKBf1suCwycY\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "createdAt": {
                          "type": "string",
                          "example": "2024-10-25T12:51:28.774Z"
                        },
                        "email": {
                          "type": "string",
                          "example": "aminrisal@gmail.com"
                        },
                        "id": {
                          "type": "string",
                          "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                        },
                        "name": {
                          "type": "string",
                          "example": "Risal Amin"
                        },
                        "password": {
                          "type": "string",
                          "example": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq"
                        },
                        "profile": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "address": {
                                "type": "string",
                                "example": "Jakarta"
                              },
                              "createdAt": {
                                "type": "string",
                                "example": "2024-10-25T12:51:28.774Z"
                              },
                              "id": {
                                "type": "string",
                                "example": "e400db97-300b-4e1b-aead-15cc7fc89882"
                              },
                              "identityNumber": {
                                "type": "string",
                                "example": "123456789"
                              },
                              "identityType": {
                                "type": "string",
                                "example": "PASSPORT"
                              },
                              "image": {
                                "type": "string",
                                "example": "https://ik.imagekit.io/ccrsxx/c093c949-6dc5-40cd-88a8-e6824e78730a_ysBQ8_ikR.jpg"
                              },
                              "updatedAt": {
                                "type": "string",
                                "example": "2024-11-07T14:53:40.825Z"
                              },
                              "userId": {
                                "type": "string",
                                "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                              }
                            }
                          },
                          "example": [
                            {
                              "address": "Jakarta",
                              "createdAt": "2024-10-25T12:51:28.774Z",
                              "id": "e400db97-300b-4e1b-aead-15cc7fc89882",
                              "identityNumber": "123456789",
                              "identityType": "PASSPORT",
                              "image": "https://ik.imagekit.io/ccrsxx/c093c949-6dc5-40cd-88a8-e6824e78730a_ysBQ8_ikR.jpg",
                              "updatedAt": "2024-11-07T14:53:40.825Z",
                              "userId": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                            }
                          ]
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2024-10-25T12:51:28.774Z"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "200 - Update current user profile": {
                    "value": {
                      "data": {
                        "createdAt": "2024-10-25T12:51:28.774Z",
                        "email": "aminrisal@gmail.com",
                        "id": "5ec59a42-64f4-437d-8477-3b3db84c60aa",
                        "name": "Risal Amin",
                        "password": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq",
                        "profile": [
                          {
                            "address": "Jakarta",
                            "createdAt": "2024-10-25T12:51:28.774Z",
                            "id": "e400db97-300b-4e1b-aead-15cc7fc89882",
                            "identityNumber": "123456789",
                            "identityType": "PASSPORT",
                            "image": "https://ik.imagekit.io/ccrsxx/c093c949-6dc5-40cd-88a8-e6824e78730a_ysBQ8_ikR.jpg",
                            "updatedAt": "2024-11-07T14:53:40.825Z",
                            "userId": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                          }
                        ],
                        "updatedAt": "2024-10-25T12:51:28.774Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 - Invalid body",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "53"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 07 Nov 2024 14:50:54 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"35-0mpTvn5Q36KIu6TSSR4OnXahZ+A\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "\"image\" must be a valid uri"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "400 - Invalid body": {
                    "value": {
                      "error": {
                        "message": "\"image\" must be a valid uri"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401 - Invalid token",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "37"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 07 Nov 2024 14:52:15 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"25-FIV/ZCaZ465tIhBUa/pSs9dwkFc\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Invalid token"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "401 - Invalid token": {
                    "value": {
                      "error": {
                        "message": "Invalid token"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}